Cofig isSet 追加 #2
