+++ srcs/config/ConfigParser.hpp
@@ -6,28 +6,7 @@
 #include <map>
 
 #include "Config.hpp"
-
-enum class E_DirectibeType
-{
-	SERVER_NAME,
-	LISTEN,
-	ERROR_PAGE,
-	CLIENT_MAX_BODY_SIZE,
-	INDEX,
-	ALLOW_METHOD,
-	REDIRECT,
-	ALIAS,
-	AUTOINDEX,
-	CGI_EXTENSION,
-	UPLOAD_FILEPATH,
-};
-
-enum class E_BlockType
-{
-	ROOT,
-	SERVER,
-	LOCATION,
-};
+#include "ConfigConstant.hpp"
 
 class ConfigParser
 {
